refactor + feat: move util functions to util.py and broaden search formatting
