+++ app/models.py
@@ -11,14 +11,16 @@ class Ytcomment(models.Model):
     username = models.CharField(max_length=100, default="", null=True)
     timestamp = models.DateTimeField(unique=False)
     video = models.CharField(max_length=100, default="", null=True)
+    user_id = models.CharField(max_length=1, null=True)
 
 
 class Fbpost(models.Model):
     page_id = models.CharField(max_length=100, default='', null=True)
     post_id = models.CharField(max_length=100, default='', null=True)
     message = models.TextField(default="", null=True)
-    pagename = models.CharField(max_length=100, default="", null=True)
+    username = models.CharField(max_length=100, default="", null=True)
     timestamp = models.DateTimeField(unique=False)
+    user_id = models.CharField(max_length=1, null=True)
 
     def __repr__(self):
         return self.post_id if self.post_id else "FBPOST"
